// src/app/models/metodo.model.ts
import { Color } from '@ionic/core';

export class Metodo {
  nome: string;
  figuracapa: string;
  figura1?: string;
  figura2?: string;
  figura3?: string;
  figura4?: string;
  apresentacao: string;
  comofazer: string;
  exemplo1: string;
  exemplo2?: string;
  color: string;

  constructor() {}
}